Type-casted the input and return values to double in test_raise_to_power

+++ Test/unit/math_utils_test.c
@@ -880,10 +880,10 @@ static void test_log2_power2(void) {
 }
 
 static void test_raise_to_power(void) {
-    TEST_ASSERT_EQ(raise_to_power(4, 3), 64, "4 to power 3");
-    TEST_ASSERT_EQ(raise_to_power(12, 5), 248832, "12 to power 5");
-    TEST_ASSERT_EQ(raise_to_power(-2, 3), -8, "-2 to power 3");
-    TEST_ASSERT_EQ(raise_to_power(-2, 4), 16, "-2 to power 4");
+    TEST_ASSERT_EQ(raise_to_power((double)4, (double)3), (double)64, "4 to power 3");
+    TEST_ASSERT_EQ(raise_to_power((double)12, (double)5), (double)248832, "12 to power 5");
+    TEST_ASSERT_EQ(raise_to_power((double)-2, (double)3), (double)-8, "-2 to power 3");
+    TEST_ASSERT_EQ(raise_to_power((double)-2, (double)4), (double)16, "-2 to power 4");
 }
 
 void run_math_utils_tests(void) {
